### 2024-11-22 #7516 @editorialbot
[PRE REVIEW]: QhX: A Python package for periodicity detection in red noise
**Submitting author:** @lionandjelka (<a href="http://orcid.org/0000-0001-5139-1978">Andjelka Kovacevic</a>)
**Repository:** https://github.com/lionandjelka/QhX1
**Branch with paper.md** (empty if default branch): paper-submission
**Version:** v0.2.0
