### 2024-11-23 #7516 @warrickball
Hi @lionandjelka, and thanks for your submission to JOSS. All our editors are currently busy so I'm going to place this on our current waiting list. In any case, the project you've submitted has a variety of issues that I've listed below that need to be addressed before we open this up for an editor to pick up, so I'm going to tag it as "paused" until the following points are addressed.

1. The repo description says it's a "test repo of Qhx". Is that still true? I see you have several different Qhx repos, so it isn't clear this is the one you intended to submit.

2. There are a lot of extra files in the repo, including at least
   * some files named `0`, `1`, `2` and `3`,
   * a bunch of CSV files,
   * some Jupyter Notebooks (which might be part of the docs),
   * probably some profiling output in `prof` and
   * some build artefacts (e.g. `Qhx.egg-info/` and `build/`).

   I don't think any of these should be in the repo and are only likely to confuse any potential reviewers.

3. The Qhx documentation describes [two versions](https://lionandjelka.github.io/QhX1/important_note.html) of the same code. A final step in the JOSS publication process will be to tag a new release version, which is difficult to reconcile with having two versions. Regardless, just for ease-of-use, it makes more sense to have a single package with appropriate options—be that different functions or relevant arguments—to select between two different modes of execution.

4. This is less important at the outset but the paper is also a bit long. The [JOSS documentation says](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain):

   > The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.

   The JOSS paper should be little more than a brief summary of the software, how it fills a current niche and how it would be (or has been) used in research. At the very least, your section describing the contents of the documentation page ("Documentation and Tutorials") can be removed.

Feel free to ask me any questions and keep me posted on your progress through these items.
