### 2024-11-23 db89881
Removed unwanted files from tracking and updated .gitignore
