+++ build/lib/QhX/output.py
@@ -1,150 +1,0 @@
-# pylint: disable=R0801
-import math
-import pandas as pd
-import numpy as np
-
-
-
-
-
-def flatten_detected_periods(detected_periods):
-    """Flatten the nested list of dictionaries and insert NaN for empty lists."""
-    flat_list = []
-    for sublist in detected_periods:
-        if sublist:  # Check if the sublist is not empty
-            flat_list.extend(sublist)
-        else:
-            # Append a dictionary with NaN values if the sublist is empty
-            flat_list.append({key: np.nan for key in ['objectid', 'sampling_i', 'sampling_j', 'period', 'upper_error', 'lower_error', 'significance', 'label']})
-    return flat_list
-
-
-def classify_periods(detected_periods):
-    """
-    Calculates IoU and compile other metrics (low errors,  upper errors, significance of detected period, and band pairs) for each quasar ID.
-    It computes Intersection Over Union (IoU) and other relevant metrics for each quasar ID based
-    on detected periods in different band pairs, while preserving NaN values.
-
-    Parameters:
-    -----------
-    detected_periods (list of dict): List of dictionaries containing detected period data.
-
-    Returns:
-    --------
-    pd.DataFrame: DataFrame containing the classification results.
-    """
-    # Flatten the list of dictionaries
-    flat_list = flatten_detected_periods(detected_periods)
-
-    # Convert flattened list to DataFrame
-    df = pd.DataFrame(flat_list)
-    def calculate_iou(radius1, radius2, distance):
-        """
-        Calculates the Intersection over Union (IoU) for two circles given their radii and the distance between their centers.
-
-        Parameters:
-        -----------
-        radius1 (float): Radius of the first circle.
-        radius2 (float): Radius of the second circle.
-        distance (float): Distance between the centers of the two circles.
-
-        Returns:
-        --------
-        float: IoU value.
-        """
-        if distance > (radius1 + radius2):
-            return 0
-        elif distance <= abs(radius1 - radius2):
-            return 1
-        else:
-            area1 = math.pi * radius1**2
-            area2 = math.pi * radius2**2
-            d = distance
-
-            # Calculate intersection area
-            part1 = math.acos((radius1**2 + d**2 - radius2**2) / (2 * radius1 * d))
-            part2 = math.acos((radius2**2 + d**2 - radius1**2) / (2 * radius2 * d))
-            intersection = part1 * radius1**2 + part2 * radius2**2 - 0.5 * (radius1**2 * math.sin(2 * part1) + radius2**2 * math.sin(2 * part2))
-
-            union = area1 + area2 - intersection
-            return intersection / union
-
-    # Initialize list to hold DataFrame rows
-    rows_list = []
-
-    # Process each unique quasar ID
-    for name in df['objectid'].unique():
-        quasar_data = df[df['objectid'] == name]
-
-        for i in range(len(quasar_data)):
-            for j in range(i + 1, len(quasar_data)):
-                row_i = quasar_data.iloc[i]
-                row_j = quasar_data.iloc[j]
-
-                # Initialize IoU and period difference
-                iou, period_diff = np.nan, np.nan
-
-                # Check if all necessary values are present to calculate IoU and period difference
-                if not pd.isna(row_i['period']) and not pd.isna(row_j['period']) and not pd.isna(row_i['upper_error']) and not pd.isna(row_i['lower_error']) and not pd.isna(row_j['upper_error']) and not pd.isna(row_j['lower_error']):
-                    # Calculate relative difference in detected periods
-                    period_diff = abs(row_i['period'] - row_j['period']) / row_i['period']
-
[…]
