+++ QhX/QhX_dynamical.egg-info/requires.txt
@@ -1,18 +1,0 @@
-numpy
-pandas
-scipy
-scikit-learn
-scikit-optimize
-libwwz
-colorednoise
-tqdm
-panel
-hvplot
-bokeh
-datashader
-pyarrow
-dask[dataframe]
-traitlets
-
-[tests]
-pytest
