### 2024-11-24 159834c
Updated paper.md to the length of <~860 words (without references) for JOSS submission
