+++ paper/paper.md
@@ -54,25 +54,22 @@ bibliography: paper.bib
 
 
 # Summary
+`QhX` is a Python package for detecting periodicity in red noise time series, developed as an in-kind contribution to the Vera C. Rubin Observatory Legacy Survey of Space and Time [LSST, @ivezic2019]. Traditional Fourier-based methods often struggle with red noise, which is common in quasar light curves and other accreting objects. `QhX` addresses these challenges with its core 2D Hybrid method [@2018MNRASK]. Input data are mapped into a time-period plane via wavelet transforms, which are (auto)correlated to produce a correlation density map in a "period-period" plane. Statistical vetting incorporates significance, upper and lower error bounds, and the novel Intersection over Union (IoU) metric to evaluate the proximity and overlap of detected periods across bands and objects. In addition to a vetted numerical catalog, `QhX` dynamically visualizes periodicity across photometric bands and objects.
 
 
-`QhX` is a Python-based package developed for periodicity detection in red noise time series data, with a focus on the Vera C. Rubin Observatory Legacy Survey of Space and Time [LSST, @ivezic2019]. Traditional methods, such as those based on the Fourier Transform, often struggle with red noise signals, which are prevalent in real-world datasets.
-The core of `QhX` is a 2D Hybrid method [cross-correlation of wavelet transforms of light curves, @2018MNRASK]  that operates in a "period-period" phase space, capable of detecting oscillations in one or more light curves. The `QhX` pipeline first transforms input data into a time-period plane via wavelets and then (auto)correlates the resulting wavelet matrices to obtain a correlation density in the period-period plane.
-After integrating the correlation density, the final decision on detected periods is made by a statistical robovetter based on the significance, upper and lower errors of detected periods, and the Intersection over Union (IoU) metric for measuring the proximity and overlap of periods across bands.
-Beyond compiling the numerical catalog of vetted periods, `QhX` offers visualization across photometric bands.
 
 # Statement of need
-![The left panel shows a 1D light curve with observational data (black error bars) and a model (blue line). `QhX` applies wavelet transforms to convert the time-series data (either observed or modeled) into the time-frequency domain, then cross-correlates wavelet transform matrices generating a correlation density map (right) in 2D peroiod-period phase space where clusters reveal consistent periodic signals. After applying statistical procedures, the package generates a numerical catalog of flagged periodic objects (bottom left) and a dynamical view of detected periods across objects (bottom right).  \label{fig:scheme}](simpleoverview.pdf)
-
-Periodic variability can be encountered across a wide range of astronomical objects, from asteroids and stars to quasars. However, identifying meaningful signals in these variable  sources is often complicated by red noise  [see, e.g., Figure 1 in @GAIA; @Kasliwal; @Kovacevic2022], which shows fractal-like patterns across time scales [@vio91; @belete18]. This, along with the non-stationary nature of signals and often unfavorable sampling [@2018Br; @2023arXiv231016896D], makes identifying coherent signals challenging. Traditional time-frequency analysis, limited by the Fourier uncertainty principle [i.e., Gabor limit, @gabor47], struggles with these complex signals, highlighting the need for a nonlinear approach [@cohen1995time; @Abry1995] for analyzing astronomical signals like quasar light curves.
-
-`QhX` provides a robust framework with features specifically designed to address these challenges.
[…]
