### 2024-11-24 63b0f39
Update .gitignore to track only paper folder
