### 2024-11-24 f37680a
Add GitHub Actions workflows in ./github/workflows
