+++ paper/paper.md
@@ -1,144 +1,0 @@
----
-title: '`QhX_dynamical`: A Python package for periodicity detection in red noise'
-tags:
-  - Python
-  - astronomy
-  - variability
-  - quasars variability
-  - binary quasars
-authors:
-  - name: Andjelka B. Kovačević
-    orcid: 0000-0001-5139-1978
-    corresponding: true
-    equal-contrib: true
-    affiliation: "1"
-  - name: Dragana Ilić
-    orcid: 0000-0002-1134-4015
-    affiliation: "1"
-  - name: Momčilo Tošić
-    equal-contrib: true
-    affiliation: "1"
-  - name: Marina Pavlović
-    orcid: 0000-0001-5560-7051
-    affiliation: "2"
-  - name: Aman Raju
-    orcid: 0000-0001-9339-0789
-    affiliation: "1"
-  - name: Mladen Nikolić
-    orcid: 0009-0002-8943-2709
-    affiliation: "1"
-  - name: Saša Simić
-    orcid: 0000-0001-7453-2016
-    affiliation: "3"
-  - name: Iva Čvorović Hajdinjak
-    orcid: 0000-0001-9208-6574
-    affiliation: "1"
-  - name: Luka Č. Popović
-    orcid: 0000-0003-2398-7664
-    affiliation: "4"
-
-affiliations:
-  - name: University of Belgrade-Faculty of Mathematics, Studentski trg 16, Belgrade, Serbia
-    index: "1"
-  - name: Mathematical Institute of Serbian Academy of Science and Arts, Serbia
-    index: "2"
-  - name: Faculty of sciences, University of Kragujevac, Radoja Domanovića 12, Serbia
-    index: "3"
-  - name: Astronomical Observatory, Belgrade, Serbia
-    index: "4"
-
-date: 14 November 2024
-bibliography: paper.bib
----
-
-
-
-# Summary
-
-
-`QhX_dynamical` is a Python-based package developed for periodicity detection in red noise time series data, with a focus on the Vera C. Rubin Observatory Legacy Survey of Space and Time [LSST, @ivezic2019]. Traditional methods, such as those based on the Fourier Transform, often struggle with red noise signals, which are prevalent in real-world datasets.
-The core of `QhX_dynamical` is a 2D Hybrid method [cross-correlation of wavelet transforms of light curves, @2018MNRASK]  that operates in a "period-period" phase space, capable of detecting oscillations in one or more light curves. The {`QhX_dynamical`} pipeline first transforms input data into a time-period plane via wavelets and then (auto)correlates the resulting wavelet matrices to obtain a correlation density in the period-period plane.
-After integrating the correlation density, the final decision on detected periods is made by a statistical robovetter based on the significance, upper and lower errors of detected periods, and the Intersection over Union (IoU) metric for measuring the proximity and overlap of periods across bands.
-Beyond compiling the numerical catalog of vetted periods, `QhX_dynamical` offers visualization across photometric bands using `QhX_dynamical`.
-
-# Statement of need
-![The left panel shows a 1D light curve with observational data (black error bars) and a model (blue line). QhX applies wavelet transforms to convert the time-series data (either observed or modeled) into the time-frequency domain, then cross-correlates wavelet transform matrices generating a correlation density map (right) in 2D peroiod-period phase space where clusters reveal consistent periodic signals. Resulting numerical  catalogs of flagged periodic objecta are obtained after integration of density maps and application of metrics.   \label{fig:scheme}](simpleoverview.pdf)
-
[…]
