+++ paper/paper.bib
@@ -1,361 +1,0 @@
-@ARTICLE{GAIA,
-       author = {{Gaia Collaboration} and {Eyer}, L. and {Rimoldini}, L. and {Audard}, M. and {Anderson}, R.~I. and {Nienartowicz}, K. and {Glass}, F. and {Marchal}, O. and {Grenon}, M. and {Mowlavi}, N. and {Holl}, B. and {Clementini}, G. and {Aerts}, C. and {Mazeh}, T. and {Evans}, D.~W. and {Szabados}, L. and {Brown}, A.~G.~A. and {Vallenari}, A. and {Prusti}, T. and {de Bruijne}, J.~H.~J. and {Babusiaux}, C. and {Bailer-Jones}, C.~A.~L. and {Biermann}, M. and {Jansen}, F. and {Jordi}, C. and {Klioner}, S.~A. and {Lammers}, U. and {Lindegren}, L. and {Luri}, X. and {Mignard}, F. and {Panem}, C. and {Pourbaix}, D. and {Randich}, S. and {Sartoretti}, P. and {Siddiqui}, H.~I. and {Soubiran}, C. and {van Leeuwen}, F. and {Walton}, N.~A. and {Arenou}, F. and {Bastian}, U. and {Cropper}, M. and {Drimmel}, R. and {Katz}, D. and {Lattanzi}, M.~G. and {Bakker}, J. and {Cacciari}, C. and {Casta{\~n}eda}, J. and {Chaoul}, L. and {Cheek}, N. and {De Angeli}, F. and {Fabricius}, C. and {Guerra}, R. and {Masana}, E. and {Messineo}, R. and {Panuzzo}, P. and {Portell}, J. and {Riello}, M. and {Seabroke}, G.~M. and {Tanga}, P. and {Th{\'e}venin}, F. and {Gracia-Abril}, G. and {Comoretto}, G. and {Garcia-Reinaldos}, M. and {Teyssier}, D. and {Altmann}, M. and {Andrae}, R. and {Bellas-Velidis}, I. and {Benson}, K. and {Berthier}, J. and {Blomme}, R. and {Burgess}, P. and {Busso}, G. and {Carry}, B. and {Cellino}, A. and {Clotet}, M. and {Creevey}, O. and {Davidson}, M. and {De Ridder}, J. and {Delchambre}, L. and {Dell'Oro}, A. and {Ducourant}, C. and {Fern{\'a}ndez-Hern{\'a}ndez}, J. and {Fouesneau}, M. and {Fr{\'e}mat}, Y. and {Galluccio}, L. and {Garc{\'\i}a-Torres}, M. and {Gonz{\'a}lez-N{\'u}{\~n}ez}, J. and {Gonz{\'a}lez-Vidal}, J.~J. and {Gosset}, E. and {Guy}, L.~P. and {Halbwachs}, J. -L. and {Hambly}, N.~C. and {Harrison}, D.~L. and {Hern{\'a}ndez}, J. and {Hestroffer}, D. and {Hodgkin}, S.~T. and {Hutton}, A. and {Jasniewicz}, G. and {Jean-Antoine-Piccolo}, A. and {Jordan}, S. and {Korn}, A.~J. and {Krone-Martins}, A. and {Lanzafame}, A.~C. and {Lebzelter}, T. and {L{\"o}ffler}, W. and {Manteiga}, M. and {Marrese}, P.~M. and {Mart{\'\i}n-Fleitas}, J.~M. and {Moitinho}, A. and {Mora}, A. and {Muinonen}, K. and {Osinde}, J. and {Pancino}, E. and {Pauwels}, T. and {Petit}, J. -M. and {Recio-Blanco}, A. and {Rich
[…]
