### 2024-11-25 b4b658c
Exclude setup.py from tracking and update other files
