+++ build/lib/QhX/algorithms/superlets/superlet.py
@@ -1,0 +1,507 @@
+# -*- coding: utf-8 -*-
+#
+# Time-frequency analysis with superlets
+# Based on 'Time-frequency super-resolution with superlets'
+# by Moca et al., 2021 Nature Communications
+#
+# Implementation by Gregor Mönke: github.com/tensionhead
+#
+
+
+import numpy as np
+from scipy.signal import fftconvolve
+
+
+def superlet(data_arr, samplerate, scales, order_max, order_min=1, c_1=3, adaptive=False):
+    """
+    Performs Superlet Transform (SLT) on time-series data.
+
+    This function computes either a multiplicative SLT or a fractional adaptive SLT (FASLT),
+    based on the 'adaptive' flag. Multiplicative SLT is recommended for analysis within a narrow
+    frequency band, while FASLT is suitable for a broader range of frequencies. The transform
+    combines multiple Morlet wavelets with varying cycles to achieve super-resolution in
+    time-frequency analysis.
+
+    Parameters
+    ----------
+    data_arr : ndarray
+        Uniformly sampled time-series data. The first dimension is treated as the time axis.
+    samplerate : float
+        Sampling rate of the time-series data in Hz.
+    scales : ndarray
+        Array of scales for the wavelet transform, ordered from high to low for `adaptive=True`.
+    order_max : int
+        Maximum order for the superlet set, dictating the highest number of cycles.
+    order_min : int, optional
+        Minimum order for the superlet set, default is 1.
+    c_1 : int, optional
+        Base number of cycles for the lowest-order Morlet wavelet, default is 3.
+    adaptive : bool, optional
+        Flag to choose between multiplicative SLT and FASLT, default is False.
+
+    Returns
+    -------
+    ndarray: Geometric mean of the complex time-frequency representation of the input data.
+
+    References
+    ----------
+
+    - `Moca, Vasile V., et al. 2021,Time-frequency super-resolution with superlets, Nat. Comm. 12.1, 1-18 <https://www.nature.com/articles/s41467-020-20539-9>`_.
+    """
+    # Function implementation...
+
+    # adaptive SLT
+    if adaptive:
+
+        gmean_spec = FASLT(data_arr, samplerate, scales, order_max, order_min, c_1)
+
+    # multiplicative SLT
+    else:
+
+        gmean_spec = multiplicativeSLT(
+            data_arr, samplerate, scales, order_max, order_min, c_1
+        )
+
+    return gmean_spec
+
+
+def multiplicativeSLT(data_arr, samplerate, scales, order_max, order_min=1, c_1=3):
+    """
+    Performs a multiplicative Superlet Transform on time-series data.
+
+    This function computes the multiplicative Superlet Transform by combining Morlet wavelets
+    with increasing cycles, resulting in a super-resolution time-frequency representation.
+
+    Parameters
+    ----------
+    data_arr : ndarray
+        Time-series data for analysis.
+    samplerate : float
+        Sampling rate of the data in Hz.
+    scales : ndarray
+        Array of scales for the wavelet transform.
+    order_max : int
+        Maximum order of the superlet set.
+    order_min : int, optional
+        Minimum order of the superlet set, default is 1.
+    c_1 : int, optional
+        Number of cycles for the base Morlet wavelet, default is 3.
+
+    Returns
+    -------
+    ndarray: Result of the multiplicative Superlet Transform.
+    """
+    # Function implementation...
+
+    dt = 1 / samplerate
+    # create the complete multiplicative set spanning
+    # order_min - order_max
[…]
