### 2024-11-25 75b6309
Restrict paper-submission branch to paper and workflows
