+++ build/lib/QhX/utils/mock_lc_old.py
@@ -1,238 +1,0 @@
-import numpy as np
-import colorednoise as cn
-
-
-def artificial_lc_sampled(mjd,t,y):
-    """ 
-    Returns a hypothetical light curve sampled in a given OpSim strategy.
-    User needs to provide a reference light curve for sampling (usually a continous
-    light curve with 1 day cadence, see LC_conti() function).
-    
-    Parameters:
-    -----------
-    mjd: np.array
-        Modified Julian Date obtained from OpSim. It is the time of each sampling of the light curve
-        during the LSST operation period in one of the filters and specified sky coordinates.
-    t: np.array
-        Days during the survey on which we had an observation for a continuous reference light curve.
-    y: np.array
-        Light curve magnitudes for continuous reference light curve.
-     
-    Returns:
-    --------
-    top: np.array
-        Days during the survey when we had an observation (sampling) in a given OpSim strategy.
-    yop: np.array
-        Light curve magnitude taken from the reference light curve on days we had an observation 
-        (sampling) in a given OpSim strategy.
-    """
-    
-    # Convert MJD to survey days
-    top=np.ceil(mjd-mjd.min())
-    
-    # Reference light curve sampling
-    yop=[]
-    for i in range(len(top)):
-        abs_vals = np.abs(t-top[i])
-        
-        # Find matching days and their index
-        bool_arr = abs_vals < 1
-        if bool_arr.sum() != 0:
-            index = (np.where(bool_arr)[0])[0]
-            yop.append(y[index])
-            
-        # Case when we don't have a match
-        elif bool_arr.sum() == 0:
-            yop.append(-999)
-    yop=np.asarray(yop)
-    
-    # Drop placeholder values (-999)
-    top = mjd - mjd.min()
-    top = top[yop!=-999]
-    yop = yop[yop!=-999]
-    
-    
-    
-    
-    return top,yop
-
-def artificial_stochastic_mock_lc(T, deltatc=1, oscillations=True, A=0.14, noise=0.00005, z=0, frame='observed'):
-    """ 
-    Generate one artificial light curve using a stochastic model based on the Damped random walk (DRW)
-    proccess. Parameters describing the model are characteristic amplitude ("logsig2" in code) and time 
-    scale of the exponentially-decaying variability ("tau" in code), both infered from physical 
-    quantities such are supermassive black hole mass and/or luminosity of the AGN. For further details
-    regarding the model see Kovačević et al. (2021) and references therein.
-    
-    Parameters:
-    -----------
-    T: int
-        Total time span of the light curve. It is recommended to generate light curves to be at least 
-        10 times longer than their characteristic timescale (Kozłowski 2017). 
-    deltatc: int, default=1
-        Cadence (or sampling rate) - time interval between two consecutive samplings of the light 
-        curve in days.
-    oscillations: bool, default=True
-        If True, light curve simulation will take an oscillatory signal into account.
-    A: float, default=0.14
-        Amplitude of the oscillatory signal in magnitudes (used only if oscillations=True).
-    noise: float, default=0.00005
-        Amount of noise to include in the light curve simulation.
-    z: float, default=0
-        Redshift.
[…]

[…]
