### 2024-11-29 #7516 @warrickball
Thanks @lionandjelka for a quick and comprehensive response! The repo now looks much easier to navigate and review. I'll note that the failure to detect the MIT license is an issue on our side and nothing to worry about.

This is ready for an editor to pick up but all of them are currently busy, so I'll add this to the waitlist until someone is available. Sorry for the delay!

In the meantime, you may want to try to add some of the missing DOIs to the BibTeX. If you get your BibTeX from the NASA ADS, it should have the DOIs. We know not everything has a DOI (e.g. books often don't) but they should be present where available.
