### 2024-12-01 a87ebc9
Update paper folder with new JATS files and updated PDF and DOI bibliography
