+++ paper/paper.bib
@@ -20,120 +20,193 @@ archivePrefix = {arXiv},
 @ARTICLE{vio91,
        author = {{Vio}, R. and {Cristiani}, S. and {Lessi}, O. and {Salvadori}, L.},
         title = "{3C 345: Is the Variability of Quasars Nonlinear?}",
-      journal = {Astrophysical Journal},
+      journal = {\apj},
+     keywords = {Astronomical Photometry, Quasars, Variability, Light Curve, Luminosity, Power Spectra, Stochastic Processes, Astrophysics, QUASARS},
          year = 1991,
         month = oct,
        volume = {380},
-        pages = {351--356}
+        pages = {351},
+          doi = {10.1086/170594},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1991ApJ...380..351V},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@article{belete18,
-    author = {Belete, A. Bewketu and Bravo, J. P. and Canto Martins, B. L. and Le\tilde{a}o, I. C. and De Araujo, J. M. and De Medeiros, J. R.},
-    title = "{Multifractality Signatures in Quasars Time Series – I. 3C 273}",
-    journal = {Monthly Notices of the Royal Astronomical Society},
-    volume = {478},
-    number = {3},
-    pages = {3976-3986},
-    year = {2018}}
-
+@ARTICLE{belete18,
+       author = {{Belete}, A. Bewketu and {Bravo}, J.~P. and {Canto Martins}, B.~L. and {Le{\~a}o}, I.~C. and {De Araujo}, J.~M. and {De Medeiros}, J.~R.},
+        title = "{Multifractality signatures in quasars time series - I. 3C 273}",
+      journal = {\mnras},
+     keywords = {methods: statistical, galaxies: active, quasars: individual: 3C 273, Astrophysics - Astrophysics of Galaxies},
+         year = 2018,
+        month = aug,
+       volume = {478},
+       number = {3},
+        pages = {3976-3986},
+          doi = {10.1093/mnras/sty1316},
+archivePrefix = {arXiv},
+       eprint = {1805.07287},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2018MNRAS.478.3976B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+	
 @article{gabor47,
 	author = {Gabor, Dennis },
 	date = {1947/05/01},
-	journal = {Nat},
+	date-added = {2024-12-01 11:21:06 +0100},
+	date-modified = {2024-12-01 11:21:06 +0100},
+	doi = {10.1038/159591a0},
+	id = {GABOR1947},
+	isbn = {1476-4687},
+	journal = {Nature},
 	number = {4044},
 	pages = {591--594},
-	title = "{Acoustical Quanta and the Theory of Hearing}",
+	title = {Acoustical Quanta and the Theory of Hearing},
+	url = {https://doi.org/10.1038/159591a0},
 	volume = {159},
-	year = {1947}}
+	year = {1947},
+	bdsk-url-1 = {https://doi.org/10.1038/159591a0}}
+
+
+
 
 @book{cohen1995time,
   title={Time-frequency Analysis},
-  author={Cohen, L},
-  publisher={Englewood Cliffs, NJ},
-  year={1995}
+  author={Cohen, L.},
+  isbn={9780135945322},
+  series={Electrical engineering signal processing},
+  url={https://books.google.rs/books?id=CSKLQgAACAAJ},
+  year={1995},
[…]
