### 2024-12-01 fab8ecd
Update paper folder with new JATS files and updated PDF and bibliography
