+++ paper/paper.bib
@@ -295,6 +295,40 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{2023arXiv231016896D,
+       author = {{D'Orazio}, Daniel J. and {Charisi}, Maria},
+        title = "{Observational Signatures of Supermassive Black Hole Binaries}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2023,
+        month = oct,
+          eid = {arXiv:2310.16896},
+        pages = {arXiv:2310.16896},
+          doi = {10.48550/arXiv.2310.16896},
+archivePrefix = {arXiv},
+       eprint = {2310.16896},
+ primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023arXiv231016896D},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{2019J,
+       author = {{Johnson}, Michael A.~C. and {Gandhi}, Poshak and {Chapman}, Adriane P. and {Moreau}, Luc and {Charles}, Philip A. and {Clarkson}, William I. and {Hill}, Adam B.},
+        title = "{Prospecting for periods with LSST - low-mass X-ray binaries as a test case}",
+      journal = {Monthly Notices of the Royal Astronomical Society},
+     keywords = {surveys, X-rays: binaries, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - High Energy Astrophysical Phenomena},
+         year = 2019,
+        month = mar,
+       volume = {484},
+       number = {1},
+        pages = {19-30},
+          doi = {10.1093/mnras/sty3466},
+archivePrefix = {arXiv},
+       eprint = {1809.09141},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019MNRAS.484...19J},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
 
 
 @article{Kasliwal,
