### 2025-05-05 #8163 @editorialbot
[REVIEW]: QhX: A Python package for periodicity detection in red noise
**Submitting author:** @lionandjelka (<a href="http://orcid.org/0000-0001-5139-1978">Andjelka Kovacevic</a>)
**Repository:** https://github.com/lionandjelka/QhX1
**Branch with paper.md** (empty if default branch): paper-submission
**Version:** 0.2.1
