### 2025-06-12 #8163 @ivastar
@danhey, @sgeorge91 pinging this thread. Have you had a chance to review the submission? Any recommendations you have can be put into issues on the QhX repository or in the a comment here. Please let me know if you have any questions.
