### 2025-08-07 #8163 @sgeorge91
Apologies for being so late with this. The paper is well written and the package is really well documented. I tried running it, and it was easy to install and the provided tests worked really well. There was only one part of the checklist that I could not tick off at this point, and that was 'Do the authors describe how this software compares to other commonly-used packages?'. Could the authors maybe discuss this more in the manuscript?
A couple of other minor queries,
1. Is there a restriction on the input data format?
2. Is the IoU metric introduced to evaluate the overlap described for the first time in this manuscript?
