### 2025-08-11 0c4a87b
Update paper folder with Reviewer changes
