+++ paper/paper.md
@@ -86,11 +86,13 @@ The package is organized as follows:
    - `correlation` function within the `utils` module supports the 2D Hybrid method by converting light curve data into wavelet matrices and performing (auto)correlation, creating correlation density.
 
 2. **Signal Detection and Validation**:
-   - `detection` module identifies candidate periodic signals and assesses their validity using statistical measures (significance and upper and lower  error [@2019J]). The Intersection over Union (IoU) metric identifies overlapping periods across bands and objects.
+   - `detection` module identifies candidate periodic signals and assesses their validity using statistical measures (significance and upper and lower  error [@2019J]). The Intersection over Union (IoU) metric identifies overlapping periods across bands and objects. To our knowledge, this is the first application of the Intersection-over-Union (IoU) metric to quantify the overlap between detected and reference periods in astronomical time-series analysis.
    - Statistical vetting categorizes detected periods for each object and band as reliable, medium, or poor.
    
 
 3. **Data Management**:
+   - `QhX` assumes input time-series data in a simple tabular format containing time, flux (or magnitude), and associated uncertainties. Examples in the documentation illustrate how to map data from other commonly used formats into this structure.
+
    - `data_manager` and `data_manager_dynamical` modules manage data flow, data loading, outlier removal, and format compatibility. 
    - `batch_processor` and `parallelization_solver` modules optimize task distribution across multiple processors.
 
@@ -104,12 +106,8 @@ The package is organized as follows:
 
 # Representative Applications
 
-The `QhX` method has been applied to:
-
--Quasar periodicity detection [@2018MNRASK; @2019ApJK; @2020OAstK; @2023AJF].
--Quasi-Periodic Oscillations detection [@2020MNRASK].
--Very Low-Frequency (VLF) signals variability in the vicinity of earthquakes [@mathKovacevic].
--`QhX` is [the LSST directable software in-kind contribution](https://www.lsst.org/scientists/international-drh-list).
[…]
