+++ paper/paper.bib
@@ -380,3 +380,63 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2024A&A...690A.365S},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
+@ARTICLE{Graham2015,
+       author = {{Graham}, Matthew J. and {Djorgovski}, S.~G. and {Stern}, Daniel and {Glikman}, Eilat and {Drake}, Andrew J. and {Mahabal}, Ashish A. and {Donalek}, Ciro and {Larson}, Steve and {Christensen}, Eric},
+        title = "{A possible close supermassive black-hole binary in a quasar with optical periodicity}",
+      journal = {\nat},
+     keywords = {Astrophysics - Astrophysics of Galaxies, Astrophysics - High Energy Astrophysical Phenomena},
+         year = 2015,
+        month = feb,
+       volume = {518},
+       number = {7537},
+        pages = {74-76},
+          doi = {10.1038/nature14143},
+archivePrefix = {arXiv},
+       eprint = {1501.01375},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2015Natur.518...74G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Zhu2020,
+       author = {{Zhu}, Xing-Jiang and {Thrane}, Eric},
+        title = "{Toward the Unambiguous Identification of Supermassive Binary Black Holes through Bayesian Inference}",
+      journal = {\apj},
+     keywords = {Supermassive black holes, Quasars, Galaxy mergers, Bayesian statistics, 1663, 1319, 608, 1900, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2020,
+        month = sep,
+       volume = {900},
+       number = {2},
+          eid = {117},
+        pages = {117},
+          doi = {10.3847/1538-4357/abac5a},
+archivePrefix = {arXiv},
+       eprint = {2004.10944},
+ primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020ApJ...900..117Z},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+@article{nina2020,
+  title={Variation in natural short-period ionospheric noise, and acoustic and gravity waves revealed by the amplitude analysis of a VLF radio signal on the occasion of the Kraljevo earthquake (Mw= 5.4)},
+  author={Nina, Aleksandra and Pulinets, Sergey and Biagi, Pier Francesco and Nico, Giovanni and Mitrovi{\'c}, Sr{\dj}an T and Radovanovi{\'c}, Milan and Popovi{\'c}, Luka {\v C}},
+  journal={Science of the Total Environment},
+  volume={710},
+  pages={136406},
+  year={2020},
+  publisher={Elsevier}
+}
+
+@article{oh2018,
+  title={Periodic oscillations in the D region ionosphere after the 2011 Tohoku earthquake using LF standard radio waves},
+  author={Ohya, Hiroyo and Tsuchiya, Fuminori and Takishita, Yuta and Shinagawa, Hiroyuki and Nozaki, Kenro and Shiokawa, Kazuo},
+  journal={Journal of Geophysical Research: Space Physics},
+  volume={123},
+  number={6},
+  pages={5261--5270},
+  year={2018},
+  publisher={Wiley Online Library}
+}
