### 2025-11-14 lionandjelka/QhX1#13 @Zstone19
Issue with Importing QhX
I'm having an issue trying to import QhX. If I place my script in the `QhX1/` directory, it will work fine, but if the script is anywhere else, it will give a `ModuleNotFoundError`. Placing the additional sub-packages in the [tool.setuptools] section of the pyproject.toml file is a possible solution.

## System Details
**OS:** Linux x86_64
**Python version:** 3.10.9
**QhX version:** 0.2.0

## Installation Method
```
git clone https://github.com/lionandjelka/QhX1.git
cd QhX1
pip install .
pip install pytest
```

## Example Code
```
import QhX
```

## Output
```
Traceback (most recent call last):
  File "/home/stone28/projects/Miscellanious/JOSS/TestingQhX/test.py", line 1, in <module>
    import QhX
  File "/home/stone28/anaconda3/envs/joss/lib/python3.10/site-packages/QhX/__init__.py", line 24, in <module>
    from QhX.calculation import get_full_width,periods, signif_johnson
  File "/home/stone28/anaconda3/envs/joss/lib/python3.10/site-packages/QhX/calculation.py", line 8, in <module>
    from QhX.algorithms.wavelets.wwtz import *
ModuleNotFoundError: No module named 'QhX.algorithms'
```
