### 2025-11-14 #8163 @Zstone19
Apologies for taking a bit longer than expected. The documentation is very in-depth and informative, and the paper is thorough and easily understood. The software was quick to install, and I was able to run the tutorials and tests smoothly. I noticed one minor issue when testing the code:

- I was not able to import QhX if the script wasn't located in the `QhX1/` directory. I've raised a corresponding [issue](https://github.com/lionandjelka/QhX1/issues/13) detailing the problem on the source repository.
