### 2025-11-16 fb7f982
Update pyproject.toml and rebuild documentation (version update + no notebooks)
