+++ source/conf.py
@@ -11,8 +11,8 @@ copyright = '2023, Andjelka Kovacevic'
 author = 'Andjelka Kovacevic, LSST SER-SAG-S1 team'
 release = '2024'
 
-version = '0.2.0'  # The short X.Y version
-release = '0.2.0'  # The full version, including alpha/beta/rc tags
+version = '0.2.1'  # The short X.Y version
+release = '0.2.1'  # The full version, including alpha/beta/rc tags
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
