### 2025-11-16 lionandjelka/QhX1#13 @lionandjelka
Hi @Zstone19,

Thank you so very much for the invaluable report and for testing the package so thoroughly.

You were completely right: in version 0.2.0 the pyproject.toml file did not  include all subpackages (such as QhX.algorithms.wavelets). Because of that, these internal modules were not included during installation with pip install , and the import only worked when the script was placed inside the QhX1/ directory.

I have now fixed this and released a new version QhX 0.2.1.
The relevant section of pyproject.toml has been updated to:

```[tool.setuptools]
include-package-data = true
[tool.setuptools.packages.find]
where = ["."]
include = ["QhX*"]
```

Could you please try:
```pip uninstall -y QhX   # optional cleanup
git clone https://github.com/lionandjelka/QhX1.git
cd QhX1
pip install .
```
Then from any directory:
`import QhX
`
Please let me know if everything works correctly on your system — if so, I will close this issue.
Thank you most kindly  again for catching this and helping improve the package! 🙏
