### 2025-12-17 1138ba8
Refine language and formatting in paper.md
Updated text for clarity and consistency, including capitalization and phrasing adjustments throughout the document.
