### 2025-12-17 4766ea2
Enhance clarity and precision in paper.md
