+++ paper/paper.md
@@ -54,21 +54,21 @@ bibliography: paper.bib
 
 
 # Summary
-`QhX` is a Python package for detecting periodicity in red noise time series, developed as an in-kind contribution to the Vera C. Rubin Observatory Legacy Survey of Space and Time [LSST, @ivezic2019]. Traditional Fourier-based methods often struggle with red noise, which is common in quasar light curves and other accreting objects. `QhX` addresses these challenges with its core 2D hybrid method [@2018MNRASK]. Input data are mapped into a time-period plane via wavelet transforms, which are (auto)correlated to produce a correlation density map in a "period-period" plane. Statistical vetting incorporates significance, upper and lower error bounds, and the novel intersection over union (IoU) metric to evaluate the proximity and overlap of detected periods across bands and objects. In addition to a vetted numerical catalog, `QhX` dynamically visualizes periodicity across photometric bands and objects.
+`QhX` is a Python package for detecting periodicity in red noise time series, developed as an in-kind contribution to the Vera C. Rubin Observatory Legacy Survey of Space and Time [LSST, @ivezic2019]. Traditional Fourier-based methods often struggle with red noise, which is common in quasar light curves and other accreting objects. `QhX` addresses these challenges with its core 2-dimensional (2D) hybrid method [@2018MNRASK]. Input data are mapped into a time-period plane via wavelet transforms, which are (auto)correlated to produce a correlation density map in a "period-period" plane. Statistical vetting incorporates significance, upper and lower error bounds, and the novel intersection over union (IoU) metric to evaluate the proximity and overlap of detected periods across bands and objects. In addition to a vetted numerical catalog, `QhX` dynamically visualizes periodicity across photometric bands and objects.
 
 
 
 # Statement of need
-![The left panel shows a 1D light curve of observational data (black error bars) and a model (blue line). `QhX` transforms the time-series data into the time-frequency domain and cross-correlates wavelet matrices to produce a 2D period-period correlation map (right), where clusters indicate periodic signals. After map integration, statistical vetting generates a numerical catalog of flagged periodic objects (bottom left) and a dynamic view of detected periods across objects and bands (bottom right). \label{fig:scheme}](simpleoverview.pdf)
+![The left panel shows a one-dimensional (1D) light curve of observational data (black error bars) and a model (blue line). `QhX` transforms the time series data into the time-frequency domain and cross-correlates wavelet matrices to produce a 2D period-period correlation map (right). Clusters in this map indicate periodic signals. After map integration, statistical vetting generates a numerical catalog of flagged periodic objects (bottom left) and a dynamic view of detected periods across objects and bands (bottom right). \label{fig:scheme}](simpleoverview.pdf)
 
 
 
-Periodic variability spans a range of astronomical objects, from asteroids to quasars. Identifying meaningful signals is complicated by red noise [see, e.g., Figure 1 in @GAIA; @Kasliwal; @Kovacevic2022], which exhibits fractal-like patterns across time scales [@vio91; @belete18]. Non-stationary signals and unfavorable sampling [@2018Br; @2023arXiv231016896D] further obscure coherent patterns. Traditional time-frequency methods, constrained by the Fourier uncertainty principle [i.e., Gabor limit, @gabor47], often fail with such complex signals, highlighting the need for nonlinear approaches [@cohen1995time; @Abry1995].
[…]
