### 2025-12-17 3b390b2
Add acknowledgements and citations for core dependencies
