+++ paper/paper.bib
@@ -439,3 +439,64 @@ archivePrefix = {arXiv},
   publisher={Wiley Online Library},
   doi = {10.1029/2018JA025289}
 }
+
+@article{numpy,
+	author = {Harris, Charles R. and Millman, K. Jarrod and van der Walt, St{\'e}fan J. and others},
+	title = {Array programming with {NumPy}},
+	journal = {Nature},
+	volume = {585},
+	pages = {357--362},
+	year = {2020},
+	doi = {10.1038/s41586-020-2649-2}
+}
+
+@article{scipy,
+	author = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and others},
+	title = {{SciPy} 1.0: Fundamental Algorithms for Scientific Computing in Python},
+	journal = {Nature Methods},
+	volume = {17},
+	pages = {261--272},
+	year = {2020},
+	doi = {10.1038/s41592-019-0686-2}
+}
+
+@misc{libwwz,
+	author = {{ISLA-UH}},
+	title = {libwwz: A Python library for the Weighted Wavelet Z-Transform},
+	howpublished = {\url{https://github.com/ISLA-UH/libwwz}},
+	year = {2024}
+}
+
+@article{holoviews,
+	author = {Stevens, Jean-Luc and Rudiger, Philipp},
+	title = {HoloViews: Building complex visualizations easily for reproducibility},
+	journal = {Journal of Open Source Software},
+	volume = {3},
+	number = {28},
+	pages = {892},
+	year = {2018},
+	doi = {10.21105/joss.00892}
+}
+
+@software{holoviews,
+	author = {Rudiger, Philipp and Stevens, Jean-Luc and Bednar, James A. and
+	Nijholt, Bas and Mease, James and Andrew, Brian C. and
+	Randelhoff, Achim and Tenner, Volker and maxalbert and
+	Kaiser, Martin and ea42gh and Samuels, Joshua and
+	stonebig and Pevey, Kevin and LB and Tolmie, Andrew and
+	Stephan, Daniel and Hoxbro, Mads and others},
+	title = {HoloViz/HoloViews: Version 1.14.2},
+	year = {2021},
+	publisher = {Zenodo},
+	version = {v1.14.2},
+	doi = {10.5281/zenodo.4581995}
+}
+
+
+@misc{bokeh,
+	author = {{Bokeh Development Team}},
+	title = {Bokeh: Python library for interactive visualization},
+	year = {2018},
+	howpublished = {\url{https://bokeh.pydata.org/en/latest/}}
+}
+
