### 2025-12-18 #8163 @lionandjelka
Hi @ivastar

Thank you so very much  for  your kind message! I confirm that I :

- Double checked authors and affiliations (including ORCIDs) and made necessary changes for coauthors that changed affiliations
- Made a release of the software with the latest changes from the review and  the version number  is  v0.2.1.
- Archived the release on Zenodo  and DOI is https://doi.org/10.5281/zenodo.17980656
- Checked  that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

- Made sure that the license listed for the archive is the same as the software license.
Thank you so very much for your efforts!
