### 2025-12-19 2eba017
Remove ISBN from paper.bib
Removed ISBN entry from the bibliography. I think this is causing the failure in submission.
