+++ paper/paper.bib
@@ -56,7 +56,6 @@ archivePrefix = {arXiv},
 	date-modified = {2024-12-01 11:21:06 +0100},
 	doi = {10.1038/159591a0},
 	id = {GABOR1947},
-	isbn = {1476-4687},
 	journal = {Nature},
 	number = {4044},
 	pages = {591--594},
