### 2025-12-19 #8163 @ivastar
@lionandjelka I just fixed an issue with an ISBN number in one of the references. Can you please merge the pull request I made?
