### 2025-12-20 12802ba
Fix duplicate holoviews BibTeX key for JOSS XML generation
