+++ paper/paper.bib
@@ -466,16 +466,7 @@ archivePrefix = {arXiv},
 	year = {2024}
 }
 
-@article{holoviews,
-	author = {Stevens, Jean-Luc and Rudiger, Philipp},
-	title = {HoloViews: Building complex visualizations easily for reproducibility},
-	journal = {Journal of Open Source Software},
-	volume = {3},
-	number = {28},
-	pages = {892},
-	year = {2018},
-	doi = {10.21105/joss.00892}
-}
+
 
 @software{holoviews,
 	author = {Rudiger, Philipp and Stevens, Jean-Luc and Bednar, James A. and
