### 2025-12-28 7f64857
Initial editorial changes for publication in JOSS
