+++ paper/paper.md
@@ -1,10 +1,10 @@
 ---
-title: '`QhX`: A Python package for periodicity detection in red noise'
+title: 'QhX: A Python package for periodicity detection in red noise'
 tags:
   - Python
   - astronomy
   - variability
-  - quasars variability
+  - quasar variability
   - binary quasars
 authors:
   - name: Andjelka B. Kovačević
@@ -30,7 +30,8 @@ authors:
   - name: Saša Simić
     orcid: 0000-0001-7453-2016
     affiliation: "4"
-  - name: Iva Čvorović Hajdinjak
+  - given-names: Iva
+    surname: Čvorović Hajdinjak
     orcid: 0000-0001-9208-6574
     affiliation: "1"
   - name: Luka Č. Popović
@@ -38,7 +39,7 @@ authors:
     affiliation: "5"
 
 affiliations:
-  - name: University of Belgrade-Faculty of Mathematics, Studentski trg 16, Belgrade, Serbia
+  - name: University of Belgrade - Faculty of Mathematics, Studentski trg 16, Belgrade, Serbia
     index: "1"
   - name: Rubin Observatory Project Oﬃce, 950 N. Cherry Ave., Tucson, AZ 85719, USA
     index: "2"
@@ -61,45 +62,41 @@ bibliography: paper.bib
 
 
 # Statement of need
-![The left panel shows a one-dimensional (1D) light curve of observational data (black error bars) and a model (blue line). `QhX` transforms the time series data into the time-frequency domain and cross-correlates wavelet matrices to produce a 2D period-period correlation map (right). Clusters in this map indicate periodic signals. After map integration, statistical vetting generates a numerical catalog of flagged periodic objects (bottom left) and a dynamic view of detected periods across objects and bands (bottom right). \label{fig:scheme}](simpleoverview.pdf)
-
-
-
 Periodic variability spans a range of astronomical objects, from asteroids to quasars. Identifying meaningful signals is complicated by red noise [see, for example, Figure 1 in @GAIA; @Kasliwal; @Kovacevic2022], which exhibits fractal-like patterns across time scales [@vio91; @belete18]. Non-stationary signals and unfavorable sampling [@2018Br; @2023arXiv231016896D] further obscure coherent patterns. Traditional time-frequency methods are often ineffective with such complex signals due to the Fourier–Gabor uncertainty limit [@gabor47], highlighting the need for nonlinear approaches [@cohen1995time; @Abry1995].
 
+![The left panel shows a one-dimensional (1D) light curve of observational data (black error bars) and a model (blue line). `QhX` transforms the time-series data into the time-frequency domain and cross-correlates wavelet matrices to produce a 2D period-period correlation map (right). Clusters in this map indicate periodic signals. After map integration, statistical vetting generates a numerical catalog of flagged periodic objects (bottom left) and a dynamic view of detected periods across objects and bands (bottom right). \label{fig:scheme}](simpleoverview.pdf)
+
 `QhX` provides features specifically designed to address these challenges.
- The first feature is its core 2D hybrid method (see Figure \ref{fig:scheme}), which is detailed in [@2018MNRASK] and inspired by 2D correlation spectroscopy [@NODA20; @10K]. By applying wavelet transforms, `QhX` maps time-series data into the time-frequency domain and (auto)correlates it, generating a period-period correlation density that enhances signal detection.
-Secondly, `QhX` introduces an intersection over union (IoU) metric, combined with standard statistical measures (significance, upper and lower error bounds), to evaluate the overlap of detected periods across bands and objects. Each period is represented as the center of an "IoU ball," with its radius reflecting relative error, calculated as the mean of the upper and lower error bounds - analogous to a circular aperture in photometry [@Saxena2024].
[…]
