### 2026-01-01 #8163 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
